added file 'nima.c'

+++ nima.c
@@ -1,7 +1,5 @@
 // in this editation we explain our code
 
-#include <stdio.h>
-#include <math.h>
 
 double max(double a,double b)
 {
@@ -14,7 +12,7 @@ double max(double a,double b)
 }
 
 
-double tool(int a,int b,int c,int d)
+double tool(double a,double b,int c,int d)
 {
     return pow(pow((a-c),2) + pow((b-d),2),0.5);
 }
@@ -26,10 +24,10 @@ double circum(int a1,int a2,int a3,int a4,int a5,int a6,int a7,int a8)
 }
 
 // this function calculates the area of the foursquare
-double are(int a1,int a2,int a3,int a4,int a5,int a6,int a7,int a8)
+double are(int a1,int a2,int a3,int a4,double a5,int a6,int a7,int a8)
 {
     double a = tool(a1,a2,a3,a4);
-    double b = tool(a3,a4,a5,a6);
+    float b = tool(a3,a4,a5,a6);
     double c = tool(a1,a2,a5,a6);
     // calculating the area of half of foursquare (a triangle) with heron formula
     double p1 = (a+b+c)/2;
